requirements.txt
requests==2.33.1
beautifulsoup4==4.14.3
lxml==6.0.2
pandas==3.0.2
python-dateutil==2.9.0.post0
gspread==6.2.1
google-auth==2.49.2
google-auth-oauthlib==1.3.1
